requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.20.0
alembic==1.13.2
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
asyncpg==0.29.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
colorlog==6.10.1
contourpy==1.3.0
cryptography==46.0.6
cycler==0.12.1
docopt==0.6.2
einops==0.8.2
exceptiongroup==1.3.1
fastapi==0.115.0
ffmpeg-python==0.2.0
filelock==3.19.1
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
google-ai-generativelanguage==0.6.9
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.0
googleapis-common-protos==1.73.1
greenlet==3.2.5
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.27.2
huggingface_hub==1.8.0
HyperPyYAML==1.2.3
idna==3.11
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.4.7
lightning==2.6.0
lightning-utilities==0.15.2
llvmlite==0.43.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
networkx==3.2.1
notion-client==2.2.1
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
openai-whisper==20240930
optuna==4.8.0
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
primePy==1.3
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.8.0
pydantic-settings==2.4.0
pydantic_core==2.20.0
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==8.3.2
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.6.0
pytorch-metric-learning==2.9.0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
scikit-learn==1.6.1
scipy==1.13.1
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
slack_sdk==3.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.35
starlette==0.38.6
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tomli==2.4.1
torch==2.4.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.4.0
torchmetrics==1.8.2
tqdm==4.67.3
typer==0.23.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.30.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
yt-dlp==2025.10.14
zipp==3.23.0
